Move abort utils into `core`
